requirements.txt.txt
pyTelegramBotAPI
pysocks
pysnmp
paramiko